\name{dl-package}
\alias{dl-package}
\alias{dl}
\docType{package}
\title{
\packageTitle{dl}
}
\description{
\packageDescription{dl}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dl}
\packageIndices{dl}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{dl}

Maintainer: \packageMaintainer{dl}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
